ai_generation/requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.1
pydantic>=2.7.0
groq>=0.11.0
